Refactor Livewire Server Advanced and Form components
